#6044 - Option to scale progress chart by tokens contained in documents
- Extract `ProgressMetric` / `ProgressWeighter` into new `inception-workload-api` module to break the `inception-search-core` -> `inception-workload` cycle
- widen `DocumentStateSnapshot` counts from `Integer` to `Long`
- fix MTAS canonical-doc resolution to sweep curation rows across all leaves before source-doc rows so docs whose curation and INITIAL_CAS rows live in different segments aren't double-counted

+++ inception/inception-log-api/src/main/java/de/tudarmstadt/ukp/inception/log/api/EventRepository.java
@@ -102,5 +102,5 @@ public interface EventRepository
     /**
      * Represents document state counts at a specific point in time.
      */
-    record DocumentStateSnapshot(Instant day, Map<SourceDocumentState, Integer> counts) {}
+    record DocumentStateSnapshot(Instant day, Map<SourceDocumentState, Long> counts) {}
 }
